change conso type to Double

+++ src/main/java/episen/pds/citizens/frontend/model/ConsoByDay.java
@@ -8,13 +8,13 @@ import java.sql.Date;
 public class ConsoByDay {
 
     private Date date;
-    private int consoday;
+    private Double consoday;
 
     public ConsoByDay() {
 
     }
 
-    public ConsoByDay(Date date, int consoday) {
+    public ConsoByDay(Date date, Double consoday) {
         this.date = date;
         this.consoday = consoday;
     }
@@ -27,11 +27,11 @@ public class ConsoByDay {
         this.date = date;
     }
 
-    public int getConsoday() {
+    public Double getConsoday() {
         return consoday;
     }
 
-    public void setConsoday(int consoday) {
+    public void setConsoday(Double consoday) {
         this.consoday = consoday;
     }
 }
